+++ hypervisor/arch/x86/guest/vcpu.c
@@ -8,7 +8,7 @@
 #include <schedule.h>
 
 #ifdef CONFIG_EFI_STUB
-#include <acrn_efi.h>
+#include <vm0_boot.h>
 extern struct efi_ctx* efi_ctx;
 #endif
