+++ hypervisor/boot/acpi.c
@@ -30,7 +30,7 @@
 #include <hypervisor.h>
 #include "acpi.h"
 #ifdef CONFIG_EFI_STUB
-#include <acrn_efi.h>
+#include <vm0_boot.h>
 #endif
 
 #define ACPI_SIG_RSDP             "RSD PTR " /* Root System Description Ptr */
