+++ hypervisor/boot/reloc.c
@@ -7,7 +7,7 @@
 #include <hypervisor.h>
 #include <reloc.h>
 #ifdef CONFIG_EFI_STUB
-#include <acrn_efi.h>
+#include <vm0_boot.h>
 #endif
 
 struct Elf64_Dyn {
