+++ hypervisor/bsp/uefi/efi/boot.c
@@ -36,7 +36,7 @@
 #include "efilinux.h"
 #include "stdlib.h"
 #include "boot.h"
-#include "acrn_efi.h"
+#include "vm0_boot.h"
 
 EFI_SYSTEM_TABLE *sys_table;
 EFI_BOOT_SERVICES *boot;
