+++ hypervisor/include/arch/x86/guest/vm0_boot.h
